chore: tidy up people and episodes
These changes make it possible to deploy
to Deno Deploy with deployment specific env
so that we don't have the project level
chicken and egg problem.

What is that problem?

Deno deploy deploys the app first, checks it works,
then deploys the deployment level envs.

We combatted this with project level envs, but
there's no API to automate t hose; so we need
to add them manually after a failed deploy.